Constrain policy parameter in 2-param constructor
Matching "any two arguments" can create weird ambiguities when you have
a struct whose first argument is a `Quantity`, _and_ an overload set
that takes either that struct or a `Quantity`, _and_ when you pass an ad
hoc instance of that struct to this overload set that you created with a
two-argument braced initializer list.

Overall, this is pretty rare; we hit just a single instance in the
entire AV codebase when we upgraded to 0.5.0.  But it's still a problem
that we can and should fix.

This commit includes tests that failed for vanilla 0.5.0, and makes them
pass.  We introduce an authoritative trait, `IsConversionRiskPolicy<T>`,
to make this easier to write.

Compile time impact measurements showed no cause for concern.

Fixes #525.