Add using directive to godbolt links (#540)
This is similar in spirit to other changes, such as using the "all
units" link, and automatically including `<iostream>`. Generally, when
using the provided godbolt link, I just want it to be as unfussy as
possible to write some units code. It's typically only for demonstration
or quick hacking purposes.

Using directives are broadly discouraged, but putting an appropriate
disclaimer here makes it easy for users who want explicit namespaces to
notice this and just delete it.

I've been meaning to do this for some time.