Add RELEASE procedures for "problem" compilers (#550)
This should help us keep more compiler versions working for more
releases.

It's also a good opportunity to clean up our compiler support page.  Now
we can discuss specific compilers in the "assumed support" tier, where
appropriate.

Additionally, it's good to emphasize front and center that our "best
effort" tier is basically equivalent to the "full support" tier, based
both on past experience and future expectations.  All I was trying to
say here was that it's _conceivable_ that we could hit an un-debuggable
problem on a compiler that we can't iterate on locally, and give
ourselves an escape hatch for that.  This is something that did happen
to me on mp-units, although that was probably because we used more
bleeding edge features in that project.

Fixes #548.  If the user needs a specific release artifact, we can
simply attach it to that issue.