Add `divide_in_common_units()` utility (#558)
This is probably what people want when they divide two same-dimension
quantities.  It is also a safe, idiomatic alternative to many if not
most use cases for `unblock_int_div`.

We update the troubleshooting guide while we're at it.

Helps #555, but I think we should add a `quotient_and_remainder` helper
too, to really nail down this idiom.

---------

Co-authored-by: Michael Hordijk <hordijk@aurora.tech>