Get clang format working with the updated toolchains (#584)
With the toolchain updates, we have a new repository name for the llvm
toolchain.  Look for the "latest" llvm toolchain available when running
clang-format.

Update our tool wrapper to run under the current directory so paths
passed in on the command line resolve correctly.

Update the bazel output to suppress more output when building tools.

Fixes #585.

---------

Co-authored-by: Chip Hogg <charles.r.hogg@gmail.com>