Improve efficiency of `mean` implementations (#527) (#590)
Turns out, the first version we launched has some extra assembly
instructions, because our sum starts at 0 instead of starting at the
first element.  If we could require C++17, of course we'd just use a
fold expression.  Instead, we'll switch to an indexed loop, and start at
`1`.

This also seems like a good opportunity to tighten up the assumption of
"at least two inputs", with a `static_assert`.  This makes it crystal
clear that starting with the 0th element is OK, as is starting the index
at `1` (because if we only _have_ one element, the loop will not
execute).

Fixes #526.