Add coverage for `//au:std:format`
We write some new simple tests for `<format>`, by adapting the test
cases that we already had for `{fmt}`.  Interestingly, these revealed
that we needed some tweaks in the case of clang 17.

We make this new test target `"manual"`, so that it won't get picked
up in our workflows by default.  Unfortunately, many toolchains that
"support C++20" don't actually support `<format>`!  So, we give it
another tag, `"requires_std_format"`, which individual workflows can use
to _explicitly opt out_ if necessary.  It turns out that clang 17 is the
only toolchain we currently have that appears to support `<format>`, so
we add the explicit opt-out everywhere else.

Fixes #508.