Disambiguate prefixed powered unit labels
We take a simple approach: any time we apply a prefix to a unit
containing a power, we first enclose that unit's label in square
brackets.  Strictly speaking, we don't _always_ need to do this: for
example, "Gi[B / in^2]" works out equivalently to "GiB / in^2".  But
this approach keeps the code simple, and the labels unambiguous, and
that's really what we care about for auto-generated labels.

The cases that were ambiguous with these ones were cases where we
applied a power to an already-prefixed unit (e.g.,
`inverse(milli(bytes))`, with label `mB^(-1)`).  After thinking about
it, I realized that a human reader will basically always assume that the
`m` applies to the `B` only, so these labels don't need to change.

Fixes #530.