Add `Impl` suffix for library-internal helpers
We've evolved a really nice idiom where the "user" facing alias gets the
natural name, and the struct that helps implement it gets called the
"Impl".  I put "user" in scare quotes because now we're fixing up the
library-internal utilities to be consistent with this new approach.

It's certainly a lot better than the old approach, where the
implementation struct got the natural name, and the user-facing alias
had the `T` suffix.  This was just a recipe for confusion, even for
experts.

There was at least one case whwere we were already using the `Impl`
suffix for something else, so we changed that instance to a `Helper`
suffix to free it up.

Helps #86.  This will unblock us from removing the `T` suffix from our
internal helpers in a future PR.